requirements.txt

streamlit
python-dotenv

langchain
langchain-openai
langchain-google-genai
langchain-community
langchain-huggingface

langserver
fastapi
uvicorn
sse-starlette

beautifulsoup4